# Add `name` to DimensionalArray