--- a/src/dimension.jl
+++ b/src/dimension.jl
@@ -69,7 +69,8 @@ Base.:(==)(dim1::AbDim, dim2::AbDim) =
 
 # AbstractArray methods where dims are the dispatch argument
 
-@inline rebuildsliced(A, data, I) = rebuild(A, data, slicedims(A, I)...; name = A.name)
+@inline rebuildsliced(A, data, I; name = A.name) =
+    rebuild(A, data, slicedims(A, I)...; name = name)
 
 Base.@propagate_inbounds Base.getindex(A::AbstractArray, dims::Vararg{<:AbDim{<:Number}}) =
     getindex(A, dims2indices(A, dims)...)