--- a/test/prettyprinting.jl
+++ b/test/prettyprinting.jl
@@ -41,3 +41,9 @@ s2 = sprint(show, F)
 G = sum(A; dims = Time)
 s3 = sprint(show, G)
 @test occursin("test", s3)
+
+# It should NOT propagate after binary operations
+B = DimensionalArray(rand(length.(d)...), d; name = "test")
+C = A + B
+s4 = sprint(show, C)
+@test !occursin("test", s4)