--- a/test/prettyprinting.jl
+++ b/test/prettyprinting.jl
@@ -32,12 +32,12 @@ s1 = sprint(show, A)
 @test occursin("test", s1)
 
 # Does it propagate after indexing?
-F = A[Time(1:4)]
+F = A[Ti(1:4)]
 s2 = sprint(show, F)
 @test occursin("test", s2)
 
 # Does it propagate after e.g. reducing operations?
-G = sum(A; dims = Time)
+G = sum(A; dims = Ti)
 s3 = sprint(show, G)
 @test occursin("test", s3)