CLoses #58

Adding the name and putting it into the print was straight forward.

Buuut, the name does not propagate through get index, e.g. `A[Time(1:4)]` does not propagate the name. Reducing operations like `sum` propagate the name.

I want to do this as simple as possible and extend as little behavior as possible. I couldn't yet find where exactly is the lowest-lowest-truly-lowest method of `getindex`. Can you please help here and point me there?