@rafaqz 2020-02-06
This is the lowest level getindex:
https://github.com/rafaqz/DimensionalData.jl/blob/master/src/array.jl#L18-L24

But you probably want to change `rebuildsliced`
https://github.com/rafaqz/DimensionalData.jl/blob/master/src/dimension.jl#L72

I will do some reorganising, I realise that it's a bit confusing where things are currently.