@Datseris 2020-02-07
OKay, now name propagates after indexing, but unfortunately it also propagates after binary operations, e.g. `C = A + B` gets the name of `A`. Could you also point to where the basic math happens?

Sorry to ask many basic questions but I've checked primitives.jl and methods.jl and I couldn't find it. `@which` also doesn't help, since 
```
julia> @which A + B
+(A::AbstractArray, B::AbstractArray) in Base at arraymath.jl:38
```