@rafaqz 2020-02-07
Yeah that stuff is pretty much magic.

+` on `AbstractArray` calls `similar` on the first array. It would be good to override these to make sure the dimensions actually match before adding.

But for now changing rebuild in similar would do the trick. https://github.com/rafaqz/DimensionalData.jl/blob/master/src/array.jl#L47-L59