@Datseris 2020-02-07
Alright @rafaqz this is good to go for review. 2 things from my side for you to check:

1. Is the way I print the name the best? "named $name" ?
2. Are there any performance regressions? A lot of functions were changed to keyword versions, and I am not sure whether this affects performance...