@rafaqz 2020-02-07
1. Maybe just `DimensionalArray "name"`. But whatever looks best.
2. That could be a problem, I'm not sure how much that compiles away at this stage. Check the performance at the end of the test against master (don't check against travis though, its wrong due to `--no-inline` for codecov accuracy)