@Datseris 2020-02-07
Master:
```
julia> @btime $A .+ $A;
  59.476 ms (11 allocations: 154.25 MiB)

julia> @btime $A .+ $A;
  64.484 ms (11 allocations: 154.25 MiB)

julia> @btime $A .+ $A;
  72.964 ms (11 allocations: 154.25 MiB)

julia> @btime $A .+ $A;
  65.184 ms (11 allocations: 154.25 MiB)

julia> @btime $A[Time(1:4)];
  17.549 ms (19 allocations: 51.42 MiB)

julia> @btime $A[Time(1:4)];
  17.594 ms (18 allocations: 51.42 MiB)
```

This PR:
```
julia> @btime $A[Time(1:4)];
  17.534 ms (22 allocations: 51.42 MiB)

julia> @btime $A[Time(1:4)];
  17.550 ms (22 allocations: 51.42 MiB)

julia> @btime $A .+ $A;
  47.932 ms (14 allocations: 154.25 MiB)

julia> @btime $A .+ $A;
  47.997 ms (14 allocations: 154.25 MiB)
```

`@btime` reports consistently "more" allocations in this PR, but the actual amount of memory, as well as spent time is identical. (it seems that addition is faster here but I blame this on my laptop)