@Datseris 2020-02-07
In the meantime, I am gettingn test errors in this part:
```julia
@testset "dimension mirroring methods" begin

    a = rand(5, 4)
    da = DimensionalArray(a, (Y((10, 20)), X(1:4)))

    cvda = cov(da; dims=X)
```
which gives
```

ERROR: ArgumentError: length of X (4) does not match size of array dimension (Base.OneTo(5))
```
I have trouble understanding how can this be related wit the current PR...

Since a week now I have been getting warnings like
```
WARNING: Method definition permutedims(Tuple{#s16, Vararg{DimensionalData.AbstractDimension{T, G, M} where M where G where T, N}} where #s16<:(DimensionalData.AbstractDimension{T, G, M} where M where G where T) where N, Tuple{#s16, Vararg{DimensionalData.AbstractDimension{T, G, M} where M where G where T, N}} where #s16<:(DimensionalData.AbstractDimension{T, G, M} where M where G where T) where N) in module DimensionalData at C:\Users\m300808\.julia\dev\DimensionalData\src\primitives.jl:25 overwritten at C:\Users\m300808\.julia\dev\DimensionalData\src\primitives.jl:28.
  ** incremental compilation may be fatally broken for this module **
```
which may be related?