@Datseris 2020-02-07
OKay, this is the benchmark file:
```julia
using DimensionalData, Test
using DimensionalData: Time, Z, @dim
using Dates: DateTime, Month
using BenchmarkTools

# define dims with both long name and Type name
@dim Lon "Longitude" "lon"
@dim Lat "Latitude" "lat"

dt = 12
dx = 20.0

timespan = DateTime(2001):Month(dt):DateTime(2011,12)
t = Time(timespan)
x = Lon(Vector(0.5:dx:359.5))
d = (x, t)
A = DimensionalArray(rand(length.(d)...), d)

println("Time to access via Tim(1:3)")
@btime $(A)[$(Time(1:3))];
println("Time to access via A.data[:, 1:3]")
@btime $(A.data)[1:3];

println("Time to do A+A")
@btime $(A) .+ $(A);

println("Time to do A.data + A.data")
@btime $(A.data) .+ $(A.data);
```

Master:
```
Time to access via Tim(1:3)
  672.308 ns (3 allocations: 624 bytes)
Time to access via A.data[:, 1:3]
  28.039 ns (1 allocation: 112 bytes)
Time to do A+A
  3.575 μs (7 allocations: 2.14 KiB)
Time to do A.data + A.data
  145.745 ns (1 allocation: 1.77 KiB)
```

This PR:
```
Time to access via Tim(1:3)
  907.107 ns (6 allocations: 720 bytes)
Time to access via A.data[:, 1:3]
  28.241 ns (1 allocation: 112 bytes)
Time to do A+A
  3.862 μs (9 allocations: 2.19 KiB)
Time to do A.data + A.data
  153.800 ns (1 allocation: 1.77 KiB)
```

Conclusion: **the keyword arguments made it slower**.

---

We will leave this open for a while, as I don't have the time/willpower to (a) understand and (b) solve this regression. I'll work on the doc PR instead so that we have it running.