@Datseris 2020-02-07
Yes of course, you have a correct point that I completely forgot: the dimensions are also allocated (which makes sense). With ranges that is not the case and indeed the indexing performance is identical. With the updated file we now get:

### Master
```
Time to access via X(1:3)
  69.774 ns (2 allocations: 432 bytes)
Time to access if X(1:3) is not interpolated constant
  66.393 ns (2 allocations: 432 bytes)
Time to just access A.data
  65.884 ns (2 allocations: 432 bytes)
Time to do A+A
  1.460 μs (2 allocations: 1.13 KiB)
Time to do A.data + A.data
  117.467 ns (1 allocation: 1.06 KiB)
```

### This PR
```
Time to access via X(1:3)
  108.244 ns (5 allocations: 608 bytes)
Time to access if X(1:3) is not interpolated constant
  107.521 ns (5 allocations: 608 bytes)
Time to just access A.data
  104.873 ns (5 allocations: 608 bytes)
Time to do A+A
  1.580 μs (6 allocations: 1.31 KiB)
Time to do A.data + A.data
  115.756 ns (1 allocation: 1.06 KiB)
```

which probably come from having keyword arguments. I remember quite fondly that keyword arguments bring performance loss if they are put into nanosecond functions.

Let's leave this open, until me or someone else finds the time to work on this (I actually don't know how to make this better, besides just remaking all keyword argument versions a 4 or 5 positional argument version instead).