@Datseris 2020-02-10
This makes the `name` a positional argument and restores performance to the previous value:
```
Original version:
N=10. Time to access via X(1:3)
  110.991 ns (5 allocations: 576 bytes)
N=100. Time to access via X(1:3)
  511.050 ns (5 allocations: 2.73 KiB)
N=1000. Time to access via X(1:3)
  3.843 μs (6 allocations: 23.75 KiB)

New version:
N=10. Time to access via X(1:3)
  65.542 ns (2 allocations: 416 bytes)
N=100. Time to access via X(1:3)
  368.778 ns (2 allocations: 2.58 KiB)
N=1000. Time to access via X(1:3)
  3.300 μs (3 allocations: 23.59 KiB)
```